Commit Branch
Optimize recaptcha and fix the sequence of register validation and recaptcha
