+++ lib/redis.js
@@ -1,3 +1,4 @@
+//C:\CPRG306\CapstoneV2\lib\redis.js
 // Import the Redis library using require instead of import
 const redis = require('redis');
