+++ lib/redisOperations.js
@@ -1,3 +1,4 @@
+//C:\CPRG306\CapstoneV2\lib\redis.js
 // Import the pre-configured Redis client from another module
 const client = require('./redis');
